backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==2.14.0
python-dotenv==1.2.1
